Rename `lib` module into `util`
