+++ src/cmd/cloud.rs
@@ -6,9 +6,9 @@ use std::sync::Arc;
 
 use crate::cfg::Configuration;
 use crate::cmd::fmt::{Formatter, Json, Kind, Short, Wide, Yaml};
-use crate::lib::types::Result;
 use crate::ovh::cloud;
 use crate::ovh::{Client, ClientConfiguration};
+use crate::util::types::Result;
 
 #[tracing::instrument]
 pub async fn list_tenants(config: Arc<Configuration>, output: &Kind) -> Result<()> {
