+++ src/ovh/cloud/loadbalancer/mod.rs
@@ -7,8 +7,8 @@ use prettytable::{Cell, Row, Table};
 use serde::{Deserialize, Serialize};
 
 use crate::cmd::fmt::{Short, Wide};
-use crate::lib::types;
 use crate::ovh::{Client, RestClient};
+use crate::util::types;
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct ConfigurationState {
